requirements.txt
psutil
pillow 